requirements.txt
httpx
py-cord
bs4
python-telegram-bot[job-queue]
python-dotenv
